+++ apps/users/services/withdrawal_service.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 from datetime import timedelta
+from typing import Any, cast
 
 from django.core.cache import cache
 from django.db import IntegrityError, transaction
@@ -16,6 +17,24 @@ from apps.users.utils.withdrawal_exceptions import (
     RecoveryPeriodExpiredError,
     WithdrawalRecordNotFoundError,
 )
+
+
+def get_email_verify_token_cache_key(email_token: str) -> str:
+    # auth_email_service.py의 verification_code()가 저장하는 토큰 캐시 키 형식
+    # email_code_{email} (인증 코드, 3분) 과 구분되는 인증 완료 토큰 캐시 (10분)
+    return f"email_verify_token_{email_token}"
+
+
+def get_recovery_token_cache(email_token: str) -> dict[str, str]:
+    # AuthPurpose에는 signup / find_password / recovery 3가지가 존재
+    # purpose 검증으로 다른 목적의 토큰이 계정 복구에 사용되지 않도록 구분
+    token_key = get_email_verify_token_cache_key(email_token)
+    cached: Any = cache.get(token_key)
+
+    if not isinstance(cached, dict) or cached.get("purpose") != AuthPurpose.RECOVERY.value:
+        raise InvalidRecoveryTokenError()
+
+    return cast(dict[str, str], cached)
 
 
 def withdraw_user(user: User, reason: str, reason_detail: str = "") -> Withdrawal:
@@ -59,19 +78,16 @@ def restore_user_by_token(email_token: str) -> None:
     :raises InvalidRecoveryTokenError: 토큰이 없거나 purpose가 recovery가 아닌 경우
     :raises DeletedUserError: 해당 이메일의 탈퇴 계정이 존재하지 않는 경우
     """
-    token_key = f"email_verify_token_{email_token}"
-    cached = cache.get(token_key)
-
-    if not cached or cached.get("purpose") != AuthPurpose.RECOVERY.value:
-        raise InvalidRecoveryTokenError()
-
+    # purpose 검증 및 캐시 데이터 조회 (Serializer에서 1차 검증 후 서비스에서 이메일 추출용으로 재조회)
+    cached = get_recovery_token_cache(email_token)
     email: str = cached["email"]
 
+    # is_active=False 조건으로 탈퇴 상태 유저만 조회
+    # 복구 완료 후 is_active=True가 되므로 같은 토큰 재사용 시 DoesNotExist → DeletedUserError로 자연 차단
+    # cache.delete()를 별도로 호출하지 않아도 DB 상태가 재사용을 막아줌 (피드백 반영 - 중복 제거)
     try:
         user = User.objects.get(email=email, is_active=False)
     except User.DoesNotExist:
         raise DeletedUserError()
 
-    if not cache.delete(token_key):  # 이미 사용된 토큰 검증
-        raise InvalidRecoveryTokenError()
     restore_user(user)
